wenet_asr_server/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
aistudio-sdk=0.1.7=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=pypi_0
av=12.0.0=pypi_0
babel=2.14.0=pypi_0
bce-python-sdk=0.9.6=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=hd590300_5
c-ares=1.28.1=hd590300_0
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
cfgv=3.4.0=pypi_0
charset-normalizer=3.3.2=pypi_0
clang-format=17.0.6=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.8.2=pypi_0
contourpy=1.2.1=pypi_0
cpplint=1.6.1=pypi_0
cycler=0.12.1=pypi_0
datasets=2.18.0=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.12.6=pypi_0
dill=0.3.4=pypi_0
distlib=0.3.8=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.110.1=pypi_0
filelock=3.13.3=pypi_0
flake8=3.8.2=pypi_0
flake8-bugbear=23.3.12=pypi_0
flake8-comprehensions=3.14.0=pypi_0
flake8-executable=2.1.3=pypi_0
flake8-pyi=20.5.0=pypi_0
flask=3.0.3=pypi_0
flask-babel=4.0.0=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.51.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
future=1.0.0=pypi_0
grpcio=1.62.1=py39h174d805_0
grpcio-tools=1.62.1=py39h3d6467e_0
h11=0.14.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.22.2=pypi_0
humanfriendly=10.0=pypi_0
identify=2.5.35=pypi_0
idna=3.6=pypi_0
ijson=3.2.3=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
iniconfig=2.0.0=pypi_0
itsdangerous=2.1.2=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
kaldiio=2.18.0=pypi_0
kiwisolver=1.4.5=pypi_0
langid=1.1.6=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libabseil=20240116.1=cxx17_h59595ed_2
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libgrpc=1.62.1=h15f2491_0
libnsl=2.0.1=hd590300_0
libprotobuf=4.25.3=h08a7969_0
libre2-11=2023.09.01=h5a48ba9_2
librosa=0.8.1=pypi_0
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
llvmlite=0.42.0=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
mccabe=0.6.1=pypi_0
mdurl=0.1.2=pypi_0
more-itertools=10.2.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.12.2=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
nodeenv=1.8.0=pypi_0
numba=0.59.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.99=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
onnx=1.16.0=pypi_0
onnxruntime=1.17.1=pypi_0
openai-whisper=20231117=pypi_0
openssl=3.2.1=hd590300_1
opt-einsum=3.3.0=pypi_0
packaging=24.0=pypi_0
paddle2onnx=1.1.0=pypi_0
paddleaudio=1.1.0=pypi_0
paddlefsl=1.1.0=pypi_0
paddlenlp=2.7.2=pypi_0
paddlepaddle=2.6.1=pypi_0
pandas=2.2.1=pypi_0
parameterized=0.9.0=pypi_0
pathos=0.2.8=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.2.0=pypi_0
pluggy=1.4.0=pypi_0
pooch=1.8.1=pypi_0
portaudio=19.6.0=h7b6447c_4
pox=0.3.4=pypi_0
ppasr=2.4.6=pypi_0
ppft=1.7.6.8=pypi_0
pre-commit=3.5.0=pypi_0
protobuf=5.26.1=pypi_0
psutil=5.9.8=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=15.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyaudio=0.2.11=py39h27cfd23_2
pybind11=2.12.0=pypi_0
pycodestyle=2.6.0=pypi_0
pycparser=2.21=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pyflakes=2.2.0=pypi_0
pygments=2.17.2=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.2=pypi_0
pytest=8.1.1=pypi_0
python=3.9.18=h0755675_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python-levenshtein=0.12.2=pypi_0
python_abi=3.9=4_cp39
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
rarfile=4.2=pypi_0
re2=2023.09.01=h7f4b329_2
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
resampy=0.4.3=pypi_0
rich=13.7.1=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
sentencepiece=0.2.0=pypi_0
seqeval=1.2.2=pypi_0
setuptools=68.2.2=py39h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soundcard=0.4.3=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
sqlite=3.41.2=h5eee18b_0
starlette=0.37.2=pypi_0
sympy=1.12=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.4.0=pypi_0
textgrid=1.6.1=pypi_0
threadpoolctl=3.4.0=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=pypi_0
tool-helpers=0.1.1=pypi_0
torch=2.2.2=pypi_0
torchaudio=2.2.2=pypi_0
tqdm=4.66.2=pypi_0
triton=2.2.0=pypi_0
typeguard=2.13.3=pypi_0
typer=0.12.1=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
virtualenv=20.25.1=pypi_0
visualdl=2.5.3=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zhconv=1.4.3=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=hd590300_5
